fix: Update buzzer to use standard tone() function for Wokwi compatibility
